split up script from commandline args
